Added ImageNodeUI and ChannelDataProcessorUI files.
These complete the split-up of GafferImage.NodeUIs, which is no longer.
