+++ python/GafferImageUI/ImageNodeUI.py
@@ -1,7 +1,6 @@
 ##########################################################################
 #
-#  Copyright (c) 2012, John Haddon. All rights reserved.
-#  Copyright (c) 2012-2014, Image Engine Design Inc. All rights reserved.
+#  Copyright (c) 2015, Image Engine Design Inc. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are
@@ -35,32 +34,44 @@
 #
 ##########################################################################
 
-from _GafferImageUI import *
+import fnmatch
 
-import DisplayUI
-from FormatPlugValueWidget import FormatPlugValueWidget
-from FilterPlugValueWidget import FilterPlugValueWidget
-from ChannelMaskPlugValueWidget import ChannelMaskPlugValueWidget
+import Gaffer
+import GafferUI
+import GafferImage
 
-import ImageReaderUI
-import ImageViewToolbar
-import ImageTransformUI
-import ConstantUI
-import ImageSwitchUI
-import OpenColorIOUI
-import ImageContextVariablesUI
-import ImageStatsUI
-import DeleteChannelsUI
-import ReformatUI
-import ObjectToImageUI
-import ClampUI
-import ImageWriterUI
-import GradeUI
-import ImageMixinBaseUI
-import ImageTimeWarpUI
-import ImageSamplerUI
-import MergeUI
-import ImageNodeUI
-import ChannelDataProcessorUI
+Gaffer.Metadata.registerNode(
 
-__import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", {}, subdirectory = "GafferImageUI" )
+	GafferImage.ImageNode,
+
+	"description",
+	"""
+	Base class for nodes which generate images.
+	""",
+
+	plugs = {
+
+		"out" : [
+
+			"description",
+			"""
+			The output image generated by this node.
+			""",
+
+		],
+
+	}
+
+)
+
+def __noduleCreator( plug ) :
+
+	if isinstance( plug, GafferImage.ImagePlug ) :
+		return GafferUI.StandardNodule( plug )
+
+	return None
+
+GafferUI.Nodule.registerNodule( GafferImage.ImageNode, fnmatch.translate( "*" ), __noduleCreator )
+GafferUI.PlugValueWidget.registerType( GafferImage.ImagePlug, None )
+
+Gaffer.Metadata.registerPlugValue( GafferImage.ImageNode, "enabled", "nodeUI:section", "Node" )
